// src/models/Card.ts
export type Card = {
  id: number;
  name: string;
  subtitle: string;
  number: string;
  aspects: string[];
  types: string[];
  cost: number;
  set: string;
  hp: number;
  power: number;
  arena: string;
  traits: string[];
  keyword: string[];
  rarity: string;
  front_art: string;
  back_art: string;
  possessed: number;
};